+++ src/core/mongo_protocol.erl
@@ -35,12 +35,17 @@
 -define(KillcursorOpcode, 2007).
 
 
--spec dbcoll(mongo:database(), mongo:collection()) -> bson:utf8().
+-spec dbcoll(database(), colldb()) -> bson:utf8().
+
 %@doc Concat db and collection name with period (.) in between
+dbcoll(Db, {undefined, Coll}) ->
+  dbcoll(Db, Coll);
+dbcoll(_, {Db, Coll}) ->
+  dbcoll(Db, Coll);
 dbcoll(Db, Coll) ->
   <<(binarize(Db))/binary, $., (binarize(Coll))/binary>>.
 
--spec put_message(mongo:database(), message(), requestid()) -> binary().
+-spec put_message(mc_worker_api:database(), message(), requestid()) -> binary().
 put_message(Db, #insert{collection = Coll, documents = Docs}, _RequestId) ->
   <<?put_header(?InsertOpcode),
   ?put_int32(0),
@@ -72,7 +77,7 @@ put_message(Db, #'query'{tailablecursor = TC, slaveok = SOK, nocursortimeout = N
   ?put_int32(Skip),
   ?put_int32(Batch),
   (bson_binary:put_document(Sel))/binary,
-  (case Proj of [] -> <<>>; _ -> bson_binary:put_document(Proj) end)/binary>>;
+  (add_proj(Proj))/binary>>;
 put_message(Db, #getmore{collection = Coll, batchsize = Batch, cursorid = Cid}, _RequestId) ->
   <<?put_header(?GetmoreOpcode),
   ?put_int32(0),
@@ -118,3 +123,12 @@ bit(true) -> 1.
 %% @private
 bool(0) -> false;
 bool(1) -> true.
+
+%% @private
+add_proj(Projector) when is_map(Projector) ->
+  case map_size(Projector) of
+    0 -> <<>>;
+    _ -> bson_binary:put_document(Projector)
+  end;
+add_proj([]) -> <<>>;
+add_proj(Other) -> bson_binary:put_document(Other).